refactor: use score-humanitec in cmd directory
