+++ internal/command/consts.go
@@ -5,7 +5,7 @@ Copyright 2020 The Apache Software Foundation
 This product includes software developed at
 The Apache Software Foundation (http://www.apache.org/).
 */
-package cmd
+package command
 
 const (
 	scoreFileDefault      = "./score.yaml"
