+++ internal/command/draft.go
@@ -5,7 +5,7 @@ Copyright 2020 The Apache Software Foundation
 This product includes software developed at
 The Apache Software Foundation (http://www.apache.org/).
 */
-package cmd
+package command
 
 import (
 	"github.com/spf13/cobra"
